One weird static field scope test
This test has code where `B::C = 1` comes before `module B`.

This now causes the `<AttachedClass>` and `<static-init>` symbols to get
entered before the `module B` has a chance to overwrite them with the
loc of the `ClassDef`.

We may want to change this in the future, but I'm `¯\_(ツ)_/¯` for now.